styling is easier to understand (i think) and is exactly centered
removed all the float lefts and put the main sentence in a single h3
tag (this could be a p tag as well, but didn't want to restyle it).
this takes advantage of the text-align center and doesn't require us to
manually set the width just right. @winners84 let me know what you
think.