only use google places results to ensure accuracy.
the problem is that the exact strings that google places returns don't
always return the same (read useful) lat/lng coordinates as the places
results. to ensure we have accurate location data we need to use the
places results.

because of this, we have some weird hacks to make sure a selection was
made from the auto complete box, b/c it doesn't have any built in
support for "always pick the first one on loose focus".

this will do for now, though the interaction could use some work.