always show the text being used for directions.
the directions are only set by selecting through the auto select box.
if you change the text and click outside without making a selection,
auto complete box would still have the old selection, which wouldn't
match the text. this is confusing, so we put the old text back if this
happens.